Add secure_binding_credentials section of the manifest
Instead of using a CredHub block in the manifest, we are moving to
`secure_binding_credentials`. We still keep the api_url in
`credhub.api_url`, which is going to be refactor in a following story.

We are also adding a `enabled` flag to activate secure binding. The
default value is false.

[#152202742]

Signed-off-by: Sarah Connor <sconnor@pivotal.io>