Increase timeout on contract tests to 30 minutes
When contract tests are running in paralel with system tests, the
contract tests tasks can be queued up. Increasing this timeout will
avoid us having to include serial groups everywhere in our pipelines.

Ideally, we would check the task state and wait for it to be not
"queued", but that would require more significant changes in our tests.

Increasing the timeout, for now, is a good compromise.

Co-authored-by: Derik Evangelista <devangelista@pivotal.io>