Use map[string]string to MaintenanceInfo.Public
When deserialising the config YAML using go-yaml, the library will, by
default, convert nested elements into a map[interface{}]interface{} (see
go-yaml/yaml#139). Later on, we try to serialise this struct into json,
which causes an error (as the built-in json lib will only convert
map[string]interfaces{}).

By allowing nested elements in our configuration, we were getting a json
marshalling failure on the brokerapi, which causes the catalog response
to be empty.

To simplify, we decided to enforce a flat structure in the
maintenance_info.public field.

[#161091710]