create-and-upload-releases now works from any loc
Running `vbox/create-and-upload-releases.sh` from vbox/ didn't work
unless you specify the ODB release path as an argument. This is a bit
counter intuitive given the `vbox/make-vbox-cf-env` does work from other
locations.

This change allows the `vbox/create-and-upload-releases.sh` to be
decoupled from a given location when an argument is not passed.