Update brokerapi submodule to point to code.cloudfoundry.org
This submodule isn't actually used but historically tracks the upstream
brokerapi and CI has historically used this to flag when a dependency is
out of date.

This pattern predated go modules, and will likely be removed in the
future. For now keeping the submodule and pointing to the current
remote repo.